修复header多个分号问题
