--- src/type/BusEvent.ts
@@ -1,9 +1,0 @@
-export class BusEvent {
-  static OPEN_IN_NEW_TAB = "open_in_new_tab"; //EventData: RequestData
-  static OPEN_IN_CURRENT_TAB = "open_in_current_tab"; //EventData: RequestData
-  static OPEN_IN_BACKGROUND_TAB = "open_in_background_tab"; //EventData: RequestData
-  static OPEN_TAB = "open_tab";
-  static ADD_HISTORY = "add_history"; //EventData: RequestData
-  static SEARCH_API = "search_api";
-  static SEARCH_HISTORY = "search_history";
-}
