--- src/type/OptionData.ts
@@ -1,4 +1,0 @@
-interface OptionData {
-  value: string;
-  label: string;
-}
