+++ src/type/Suggestion.ts
@@ -30,7 +30,7 @@ export function createSuggestions(range: IRange): monaco.languages.CompletionIte
   return header.map(value => createItem(value, range, value + ":"));
 }
 
-function createItem(label: string, range: IRange, insertText?: string) {
+function createItem(label: string, range: IRange, insertText?: string): monaco.languages.CompletionItem {
   return {
     label: label,
     kind: monaco.languages.CompletionItemKind.Field,
